Cirrus: Overhaul/Simplify env. var setup

+++ contrib/cirrus/setup_environment.sh
@@ -4,11 +4,22 @@ set -e
 
 source $(dirname $0)/lib.sh
 
-req_env_var USER HOME ENVLIB SCRIPT_BASE CIRRUS_BUILD_ID
+req_env_var USER HOME GOSRC SCRIPT_BASE SETUP_MARKER_FILEPATH
 
+# Ensure this script only executes successfully once and always logs ending timestamp
+[[ ! -e "$SETUP_MARKER_FILEPATH" ]] || exit 0
+exithandler() {
+    RET=$?
+    set +e
+    show_env_vars
+    record_timestamp "env. setup end"
+    echo "$(basename $0) exit status: $RET"
+    [[ "$RET" -eq "0" ]] && date +%s >> "SETUP_MARKER_FILEPATH"
+}
+trap exithandler EXIT
+
+# Must be bash, always bash
 [[ "$SHELL" =~ "bash" ]] || chsh -s /bin/bash
-
-cd "$CIRRUS_WORKING_DIR"  # for clarity of initial conditions
 
 # Verify basic dependencies
 for depbin in go rsync unzip sha256sum curl make python3 git
@@ -19,61 +30,48 @@ do
     fi
 done
 
-# Setup env. vars common to all tasks/scripts/platforms and
-# ensure they return for every following script execution.
-MARK="# Added by $0, manual changes will be lost."
-touch "$HOME/$ENVLIB"
-if ! grep -q "$MARK" "$HOME/$ENVLIB"
-then
-    cp "$HOME/$ENVLIB" "$HOME/${ENVLIB}_original"
-    # N/B: Single-quote items evaluated every time, double-quotes only once (right now).
-    for envstr in \
-        "$MARK" \
-        "export EPOCH_TEST_COMMIT=\"$CIRRUS_BASE_SHA\"" \
-        "export HEAD=\"$CIRRUS_CHANGE_IN_REPO\"" \
-        "export TRAVIS=\"1\"" \
-        "export GOSRC=\"$CIRRUS_WORKING_DIR\"" \
-        "export OS_RELEASE_ID=\"$(os_release_id)\"" \
-        "export OS_RELEASE_VER=\"$(os_release_ver)\"" \
-        "export OS_REL_VER=\"$(os_release_id)-$(os_release_ver)\"" \
-        "export TEST_REMOTE_CLIENT=\"$TEST_REMOTE_CLIENT\"" \
-        "export GOPATH=\"/var/tmp/go\"" \
-        'export PATH="$HOME/bin:$GOPATH/bin:/usr/local/bin:$PATH"' \
-        'export LD_LIBRARY_PATH="/usr/local/lib${LD_LIBRARY_PATH:+:${LD_LIBRARY_PATH}}"'
-    do
-        # Make permanent in later shells, and set in current shell
-        X=$(echo "$envstr" | tee -a "$HOME/$ENVLIB") && eval "$X" && echo "$X"
-    done
+# Sometimes environment setup needs to vary between distros
+# Note: This should only be used for environment variables, and temporary workarounds.
+#       Anything externally dependent, should be made fixed-in-time by adding to
+#       contrib/cirrus/packer/*_setup.sh to be incorporated into VM cache-images
+#       (see docs).
+case "${OS_REL_VER}" in
+    ubuntu-18) ;;
+    fedora-29) ;;
+    fedora-28) ;;
+    centos-7) ;;
+    rhel-7) ;;
+    *) bad_os_id_ver ;;
+esac
 
-    # Some environment setup needs to vary between distros
-    # Note: This should only be used for environment variables, and minor details.
-    #       Anything that could vary from one run to the next, should go into
-    #       contrib/cirrus/packer/*_setup.sh and be incorporated into VM cache-images
-    #       (see docs)
-    case "${OS_RELEASE_ID}-${OS_RELEASE_VER}" in
-        ubuntu-18) ;;
-        fedora-29) ;;
-        fedora-28) ;;
-        centos-7) ;;
-        rhel-7) ;;
-        *) bad_os_id_ver ;;
-    esac
+cd "${GOSRC}/"
+# Reload to incorporate any changes from above
+source "$SCRIPT_BASE/lib.sh"
 
-    cd "${GOSRC}/"
-    # Reload to incorporate any changes from above
-    source "$SCRIPT_BASE/lib.sh"
+echo "Installing cni config, policy and registry config"
+req_env_var GOSRC
+sudo install -D -m 755 $GOSRC/cni/87-podman-bridge.conflist \
+                       /etc/cni/net.d/87-podman-bridge.conflist
+sudo install -D -m 755 $GOSRC/test/policy.json \
+                       /etc/containers/policy.json
+sudo install -D -m 755 $GOSRC/test/registries.conf \
+                       /etc/containers/registries.conf
+# cri-o if installed will mess with testing in non-obvious ways
+rm -f /etc/cni/net.d/*cri*
 
-    case "$SPECIALMODE" in
-        rootless)
-            X=$(echo "export ROOTLESS_USER='some${RANDOM}dude'" | \
-                tee -a "$HOME/$ENVLIB") && eval "$X" && echo "$X"
-            setup_rootless
-            ;;
-        in_podman)  # Assumed to be Fedora
-            dnf install -y podman buildah
-            $SCRIPT_BASE/setup_container_environment.sh
-            ;;
-    esac
-fi
+go get github.com/onsi/ginkgo/ginkgo
+go get github.com/onsi/gomega/...
 
-show_env_vars
+case "$SPECIALMODE" in
+    rootless)
+        X=$(echo "export ROOTLESS_USER='some${RANDOM}dude'" | \
+            tee -a "$HOME/$ENVLIB") && eval "$X" && echo "$X"
+        X=$(echo "export SPECIALMODE='$SPECIALMODE'"| \
+            tee -a "$HOME/$ENVLIB") && eval "$X" && echo "$X"
+        setup_rootless
+        ;;
+    in_podman)  # Assumed to be Fedora
+        dnf install -y podman buildah
+        $SCRIPT_BASE/setup_container_environment.sh
+        ;;
+esac
